+++ scripts/compile_requirements.sh
@@ -12,9 +12,9 @@ pip-compile examples/requirements/deployment.in "$@"
 echo "dev.in"
 pip-compile examples/requirements/dev.in "$@"
 
-echo "django_2_2.in"
-pip-compile examples/requirements/django_2_2.in "$@"
-
+#echo "django_2_2.in"
+#pip-compile examples/requirements/django_2_2.in "$@"
+#
 #echo "django_3_0.in"
 #pip-compile examples/requirements/django_3_0.in "$@"
 #
@@ -24,8 +24,8 @@ pip-compile examples/requirements/django_2_2.in "$@"
 echo "django_3_2.in"
 pip-compile examples/requirements/django_3_2.in "$@"
 
-echo "django_4_0.in"
-pip-compile examples/requirements/django_4_0.in "$@"
+#echo "django_4_0.in"
+#pip-compile examples/requirements/django_4_0.in "$@"
 
 echo "django_4_1.in"
 pip-compile examples/requirements/django_4_1.in "$@"
@@ -33,8 +33,8 @@ pip-compile examples/requirements/django_4_1.in "$@"
 echo "django_4_2.in"
 pip-compile examples/requirements/django_4_2.in "$@"
 
-echo "django_2_2_and_flask.in"
-pip-compile examples/requirements/django_2_2_and_flask.in "$@"
+#echo "django_2_2_and_flask.in"
+#pip-compile examples/requirements/django_2_2_and_flask.in "$@"
 
 #echo "django_3_0_and_flask.in"
 #pip-compile examples/requirements/django_3_0_and_flask.in "$@"
@@ -45,8 +45,8 @@ pip-compile examples/requirements/django_2_2_and_flask.in "$@"
 echo "django_3_2_and_flask.in"
 pip-compile examples/requirements/django_3_2_and_flask.in "$@"
 
-echo "django_4_0_and_flask.in"
-pip-compile examples/requirements/django_4_0_and_flask.in "$@"
+#echo "django_4_0_and_flask.in"
+#pip-compile examples/requirements/django_4_0_and_flask.in "$@"
 
 echo "django_4_1_and_flask.in"
 pip-compile examples/requirements/django_4_1_and_flask.in "$@"